Completed front and backend. All is working.

+++ client/src/components/WorkoutDetails/WorkoutDetails.jsx
@@ -1,5 +1,7 @@
 import "./WorkoutDetails.css"
 import { useWorkoutsContext } from "../../hooks/useWorkoutContext"
+import { useAuthContext } from "../../hooks/useAuthContext"
+
 
 // Date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
@@ -7,10 +9,18 @@ import formatDistanceToNow from "date-fns/formatDistanceToNow"
 export default function WorkoutDetails(props) {
 
     const {dispatch} = useWorkoutsContext()
+    const {user} = useAuthContext()
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch("http://localhost:4000/api/workouts/" + props.workout._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers:  {
+                "Authorization": `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
